Remove hand-written GraphQL types, update imports, rename enums
