+++ packages/eas-cli/src/credentials/ios/api/graphql/mutations/IosAppCredentialsMutation.ts
@@ -1,10 +1,8 @@
 import gql from 'graphql-tag';
 
 import { graphqlClient, withErrorHandlingAsync } from '../../../../../graphql/client';
-import {
-  IosAppCredentials,
-  IosAppCredentialsFragment,
-} from '../../../../../graphql/types/credentials/IosAppCredentials';
+import { IosAppCredentials } from '../../../../../graphql/generated';
+import { IosAppCredentialsFragment } from '../../../../../graphql/types/credentials/IosAppCredentials';
 
 const IosAppCredentialsMutation = {
   async createIosAppCredentialsAsync(
