+++ packages/eas-cli/src/credentials/ios/api/graphql/mutations/__mocks__/AppleDeviceMutation.ts
@@ -1,8 +1,8 @@
-import { AppleDevice } from '../../../../../../graphql/types/credentials/AppleDevice';
+import { AppleDevice } from '../../../../../../graphql/generated';
 
 const AppleDeviceMutation = {
   createAppleDeviceAsync: jest.fn().mockImplementation(() => {
-    const appleDevice: AppleDevice = {
+    const appleDevice: Pick<AppleDevice, 'id' | 'identifier'> = {
       id: 'apple-device-id',
       identifier: '00009999-000D6666146B888E',
     };
