+++ packages/eas-cli/src/credentials/ios/api/graphql/queries/AppQuery.ts
@@ -1,7 +1,8 @@
 import gql from 'graphql-tag';
 
 import { graphqlClient, withErrorHandlingAsync } from '../../../../../graphql/client';
-import { App, AppFragment } from '../../../../../graphql/types/App';
+import { App } from '../../../../../graphql/generated';
+import { AppFragment } from '../../../../../graphql/types/App';
 
 const AppQuery = {
   async byFullNameAsync(fullName: string): Promise<App> {
