+++ packages/eas-cli/src/devices/actions/create/__tests__/action-test.ts
@@ -1,7 +1,7 @@
 import prompts from 'prompts';
 
 import { asMock } from '../../../../__tests__/utils';
-import { AppleTeam } from '../../../../graphql/types/credentials/AppleTeam';
+import { AppleTeam } from '../../../../graphql/generated';
 import { Account } from '../../../../user/Account';
 import DeviceCreateAction, { RegistrationMethod } from '../action';
 import { runInputMethodAsync } from '../inputMethod';
