+++ packages/eas-cli/src/devices/actions/create/__tests__/inputMethod-test.ts
@@ -2,8 +2,7 @@ import prompts from 'prompts';
 
 import { asMock } from '../../../../__tests__/utils';
 import { AppleDeviceMutation } from '../../../../credentials/ios/api/graphql/mutations/AppleDeviceMutation';
-import { AppleDeviceClass } from '../../../../graphql/types/credentials/AppleDevice';
-import { AppleTeam } from '../../../../graphql/types/credentials/AppleTeam';
+import { AppleDeviceClass, AppleTeam } from '../../../../graphql/generated';
 import { runInputMethodAsync } from '../inputMethod';
 
 jest.mock('prompts');
@@ -30,9 +29,9 @@ beforeEach(() => {
 
 describe(runInputMethodAsync, () => {
   it('should allow for multiple device registration', async () => {
-    mockDeviceData('00001111-001122334455662E', 'my iPhone', AppleDeviceClass.IPHONE);
+    mockDeviceData('00001111-001122334455662E', 'my iPhone', AppleDeviceClass.Iphone);
     asMock(prompts).mockImplementationOnce(() => ({ value: true }));
-    mockDeviceData('b12cba9856d89c932ab7a4b813c4d932534e1679', 'my iPad', AppleDeviceClass.IPAD);
+    mockDeviceData('b12cba9856d89c932ab7a4b813c4d932534e1679', 'my iPad', AppleDeviceClass.Ipad);
     asMock(prompts).mockImplementationOnce(() => ({ value: false }));
 
     const accountId = 'account-id';
