+++ packages/eas-cli/src/devices/manager.ts
@@ -3,7 +3,7 @@ import chalk from 'chalk';
 
 import { AppleTeamMutation } from '../credentials/ios/api/graphql/mutations/AppleTeamMutation';
 import { AppleTeamQuery } from '../credentials/ios/api/graphql/queries/AppleTeamQuery';
-import { AppleTeam } from '../graphql/types/credentials/AppleTeam';
+import { AppleTeam } from '../graphql/generated';
 import log from '../log';
 import { getProjectAccountNameAsync } from '../project/projectUtils';
 import { Choice, confirmAsync, promptAsync } from '../prompts';
