--- packages/eas-cli/src/graphql/types/ActivityTimelineProjectActivity.ts
@@ -1,7 +1,0 @@
-import { Actor } from './Actor';
-
-export interface ActivityTimelineProjectActivity {
-  id: string;
-  actor: Actor;
-  activityTimestamp: Date;
-}
