+++ packages/eas-cli/src/graphql/types/App.ts
@@ -1,10 +1,4 @@
 import { Fragment } from '../fragment';
-import { IosAppCredentials } from './credentials/IosAppCredentials';
-
-export interface App {
-  id: string;
-  iosAppCredentials: IosAppCredentials[];
-}
 
 export const AppFragment: Fragment = {
   name: 'app',
