--- packages/eas-cli/src/graphql/types/Project.ts
@@ -1,13 +1,0 @@
-export interface Project {
-  id: string;
-  name: string;
-  fullName: string;
-  description: string;
-  slug: string;
-  username: string;
-  published: boolean;
-  updated: Date;
-
-  /** @deprecated No longer supported */
-  iconUrl?: string;
-}
