+++ packages/eas-cli/src/graphql/types/credentials/AppleAppIdentifier.ts
@@ -1,10 +1,5 @@
 import { Fragment } from '../../fragment';
 
-export interface AppleAppIdentifier {
-  id: string;
-  bundleIdentifier: string;
-}
-
 export const AppleAppIdentifierFragment: Fragment = {
   name: 'appleAppIdentifier',
   definition: `
