+++ packages/eas-cli/src/graphql/types/credentials/AppleDeviceRegistrationRequest.ts
@@ -1,9 +1,5 @@
 import { Fragment } from '../../fragment';
 
-export interface AppleDeviceRegistrationRequest {
-  id: string;
-}
-
 export const AppleDeviceRegistrationRequestFragment: Fragment = {
   name: 'appleDeviceRegistrationRequest',
   definition: `
