+++ packages/eas-cli/src/graphql/types/credentials/AppleDistributionCertificate.ts
@@ -1,16 +1,4 @@
 import { Fragment } from '../../fragment';
-import { AppleTeam } from './AppleTeam';
-
-export interface AppleDistributionCertificate {
-  id: string;
-  certificateP12: string;
-  certificatePassword: string;
-  serialNumber: string;
-  developerPortalIdentifier: string;
-  validityNotBefore: Date;
-  validityNotAfter: Date;
-  appleTeam?: AppleTeam;
-}
 
 export const AppleDistributionCertificateFragment: Fragment = {
   name: 'appleDistCert',
