+++ packages/eas-cli/src/graphql/types/credentials/AppleProvisioningProfile.ts
@@ -1,17 +1,4 @@
 import { Fragment } from '../../fragment';
-import { AppleAppIdentifier } from './AppleAppIdentifier';
-import { AppleDevice } from './AppleDevice';
-import { AppleTeam } from './AppleTeam';
-
-export interface AppleProvisioningProfile {
-  id: string;
-  expiration: Date;
-  developerPortalIdentifier: string;
-  provisioningProfile: string;
-  appleTeam?: AppleTeam;
-  appleAppIdentifier?: AppleAppIdentifier;
-  appleDevices?: AppleDevice[];
-}
 
 export const AppleProvisioningProfileFragment: Fragment = {
   name: 'appleProvisioningProfile',
