+++ packages/eas-cli/src/user/Account.ts
@@ -1,4 +1,4 @@
-import { Account as GraphQLAccount } from '../graphql/types/Account';
+import { Account as GraphQLAccount } from '../graphql/generated';
 
 export type Account = Pick<GraphQLAccount, 'id' | 'name'>;
